Use smaller images for thumbnails if possible
This introduces a new thumb-src include which is smart about
pinterest image URLs, replacing /originals/ with one of the
downscaled URLs, and changing the extension to .jpg as required.

It also introduces a new image-thumb front matter field, so that
any page can explicitly override its small-sized image URL.

The thumb and gallery includes now use thumb-src, and gallery
prefers the image-thumb front matter field when it's available.